ENH: SIFT and writing mask of vector.
Two major changes for this commit:  Initial commit of PatchCorrelation,
which computes the orientation-invariant correlation between two
patches.  This function has not yet been debugged, so it is not yet
recommended for use.

In addition, initial commit of GenerateMaskImageFromPatch, which takes a
vector or eigenvector image and generates a mask from scratch.  This is
necessary to be able to store eigenvectors and other latent descriptors
as images and not only as vectors.  This function does work correctly
and has been debugged.